backend/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
cryptography==41.0.7

# OCR & Document Processing
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
pandas==2.1.3
numpy==1.26.2

# NLP/AI (for Week 5-6)
# openai==1.3.5  # Optional - paid API
transformers==4.35.2
torch>=2.2.0
sentence-transformers==2.2.2
joblib==1.3.2  # For ML model serialization

# Data Processing
kagglehub==0.1.0  # For downloading Kaggle datasets
opencv-python==4.8.1.78

# Export & Visualization
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
